Title: 커지는 간선의 값, Time: 99ms, Memory: 9MB, Status: Wrong Answer - Codetree

+++ 250626/커지는 간선의 값/growing-edge-value.cpp
@@ -40,7 +40,7 @@ int main()
 	return 0;
 }
 void solve() {
-	int n, m, k;
+	ll n, m, k;
 	cin >> n >> m >> k;
 
 	int a, b, c;
